test validate and try_wget
